+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.1
+
+Documented optional `exp` claim for expiration
+
+## 1.2.0
+
+Specified Health Card revocation
+
 ## 1.1.1
 
 Added verifier guidance to ignore unrecognized VC types